ensure the qty, price and weight data is sent as floats
- fixes a bug where the request may not be accepted by the Shippit API

+++ Model/Request/Order.php
@@ -657,10 +657,10 @@ class Order extends \Magento\Framework\Model\AbstractModel implements OrderInter
         $newParcel = [
             'sku' => $sku,
             'title' => $title,
-            'qty' => $qty,
-            'price' => $price,
+            'qty' => (float) $qty,
+            'price' => (float) $price,
             // if a 0 weight is provided, stub the weight to 0.2kg
-            'weight' => ($weight == 0 ? 0.2 : $weight),
+            'weight' => (float) ($weight == 0 ? 0.2 : $weight),
             'location' => $location
         ];
